+++ service/service-system/src/main/java/org/github/foxnic/web/system/service/impl/ProfileServiceImpl.java
@@ -1,33 +1,29 @@
 package org.github.foxnic.web.system.service.impl;
 
 
-import javax.annotation.Resource;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-
-import org.github.foxnic.web.domain.system.Profile;
-import org.github.foxnic.web.domain.system.ProfileVO;
-import java.util.List;
-import com.github.foxnic.api.transter.Result;
-import com.github.foxnic.dao.data.PagedList;
-import com.github.foxnic.dao.entity.SuperService;
-import com.github.foxnic.dao.spec.DAO;
-import java.lang.reflect.Field;
-import com.github.foxnic.commons.busi.id.IDGenerator;
-import com.github.foxnic.sql.expr.ConditionExpr;
 import com.github.foxnic.api.error.ErrorDesc;
+import com.github.foxnic.api.transter.Result;
+import com.github.foxnic.commons.busi.id.IDGenerator;
+import com.github.foxnic.dao.data.PagedList;
+import com.github.foxnic.dao.data.SaveMode;
+import com.github.foxnic.dao.entity.SuperService;
+import com.github.foxnic.dao.excel.ExcelStructure;
 import com.github.foxnic.dao.excel.ExcelWriter;
 import com.github.foxnic.dao.excel.ValidateResult;
-import com.github.foxnic.dao.excel.ExcelStructure;
-import java.io.InputStream;
+import com.github.foxnic.dao.spec.DAO;
+import com.github.foxnic.sql.expr.ConditionExpr;
 import com.github.foxnic.sql.meta.DBField;
-import com.github.foxnic.dao.data.SaveMode;
-import com.github.foxnic.dao.meta.DBColumnMeta;
-import com.github.foxnic.sql.expr.Select;
-import java.util.ArrayList;
-import org.github.foxnic.web.system.service.IProfileService;
+import org.github.foxnic.web.domain.system.Profile;
 import org.github.foxnic.web.framework.dao.DBConfigs;
+import org.github.foxnic.web.system.service.IConfigService;
+import org.github.foxnic.web.system.service.IProfileService;
+import org.springframework.stereotype.Service;
+
+import javax.annotation.Resource;
+import java.io.InputStream;
+import java.lang.reflect.Field;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * <p>
@@ -35,6 +31,7 @@ import org.github.foxnic.web.framework.dao.DBConfigs;
  * </p>
  * @author 李方捷 , leefangjie@qq.com
  * @since 2021-12-03 13:54:38
+ * @version
 */
 
 
@@ -44,7 +41,7 @@ public class ProfileServiceImpl extends SuperService<Profile> implements IProfil
 	/**
 	 * 注入DAO对象
 	 * */
-	@Resource(name=DBConfigs.PRIMARY_DAO) 
+	@Resource(name=DBConfigs.PRIMARY_DAO)
 	private DAO dao=null;
 
 	/**
@@ -92,7 +89,7 @@ public class ProfileServiceImpl extends SuperService<Profile> implements IProfil
 		return super.insertList(profileList);
 	}
 
-	
+
 	/**
 	 * 按主键删除 sys_profile
 	 *
@@ -103,6 +100,9 @@ public class ProfileServiceImpl extends SuperService<Profile> implements IProfil
 		Profile profile = new Profile();
 		if(id==null) return ErrorDesc.failure().message("id 不允许为 null 。");
 		profile.setId(id);
+		if(IConfigService.DEFAULT_PROFILE_ID.equals(id)) {
+			return ErrorDesc.failure().message("不允许删除默认的 Profile");
+		}
 		try {
 			boolean suc = dao.deleteEntity(profile);
 			return suc?ErrorDesc.success():ErrorDesc.failure();
@@ -149,7 +149,7 @@ public class ProfileServiceImpl extends SuperService<Profile> implements IProfil
 		return super.updateList(profileList , mode);
 	}
 
-	
+
 	/**
 	 * 按主键更新字段 sys_profile
 	 *
@@ -163,7 +163,7 @@ public class ProfileServiceImpl extends SuperService<Profile> implements IProfil
 		return suc>0;
 	}
 
-	
+
 	/**
 	 * 按主键获取 sys_profile
 	 *
